integration: add test/integration package
This commit creates a package that contains a single function that skips
tests if the test binary has been built with the `integration` build
tag.

The benefit this approach has over adding a build tag directly to the
source file in a given package is that the tests are now always built,
so a `go test` will build the code to do the integration test and run
non-integration tests.