vulnstore: implement active gc
previous to this commit GC was a passive process each updater performed
on their own.

this lead to the scenario where removing an updater from a deployed
Clair environment meant that updater's vulnerabilties were never garbage
collected.

this pr implements gc as an active process allowing Clair to run gc on a
cadence along with letting administrators trigger it.