updates: add LockSource interface
This change abstracts over how locks are created at runtime. In the
creation of the updates.Manager abstraction, the distributed use case
was the only one considered, and so the database dependency was baked
in.

This will allow the same code to be used for offline use cases, as well.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>