indexer: filter scanners during manifest check
This commit adds a filtering step in the CheckManifest state.
Now, if a manifest reports being analyzed by a particular scanner that
scanner will be omitted from all subsequent states.

This is safe to do, if a manifest was analyzed by a particular scanner,
this also means all it's layers were analyzed by the particular scanner.
Thus, the scanner need not be included in the rest of the indexing
process.

This fixes a unique constraint issue in the IndexFinished state.
Previous to this commit, an attempt to persist duplicate scanner ids was
made in the aforementioned state.

Signed-off-by: ldelossa <ldelossa@redhat.com>