all: use GOMAXPROCS instead of NumCPU
When running directly on an OS, these values will be the same. But in
containerized environments, the number of CPUs reported is the number on
the host, not the number allocated. Using the GOMAXPROCS means the
container orchestrator can configure things such that the go runtime is
less likely to constantly blow its cpu quota.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>