rpm: special-case symlinks
The behavior of chmod(2) is more subtle than it really should be.

During our permissions fixup pass, if the target of a symlink was in a
directory not fixed yet (or in system's root) calling chmod on it will
attempt to change the mode bits on a file the process won't have
permissions for. The amount of corner cases in "remove this directory
and its contents" is annoying.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>