controller: cleanups
This change removes the package-level "startState." It was used for
testing, but the "setState" method achieves the same thing.

The mutex inside the Controller struct is removed, as it was only used
in "getState" and was never used for writing. With the refactor of the
"run" method, the "getState" method was unused and also removed.

One of the tests had a copy-paste error that resulted in an unhelpful
log message.

Various automatic code simplifications were also applied.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>