dpkg: update identification method of dpkg DB dir (#456)
The current way to identify a dpkg DB can be too
exclusive in some scenarios and the "available" dir
does not always exist. This change checks the info
dir is there. This could result in some more false
positives given how generic the names are but this
is highly unlikely to lead to false positives due
to DB file format we expect.

Signed-off-by: crozzy <joseph.crosland@gmail.com>