rpm: handle cross-layer hardlinks when extracting (#475)
Logically, cross-layer links are allowed, although we hadn't run across
them in testing.

Because layers are defined to be applied in-order, a container runtime
-- which never extracts layers in isolation -- never runs into this
issue and rightly errors out when attempting to link to a nonexistent
file. Clair's use case means the program should link to an empty file,
which needs to exist on the same filesystem. The file being empty should
be okay, because the contents will be seen in a different index.

There may be an edge case where the layer dodges all the indexers by not
having the "right" name in a lower layer and linking it in a higher
layer. I don't think this is avoidable.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>
Backports: #465
(cherry picked from commit 64f62498c29e314f41bf01d3f85e3dfc378eadeb)