dockerfile: fix terminal expansion with too-small destination
This fixes a corner-case where the transformer would consume some input,
begin expansion, then need to signal for more space and lose track of
what it had consumed. I added an explicit state to signal that the
transformer should immediately try to expand, and then modified the
return path of the corner case to use that state.

Closes: PROJQUAY-2765
Signed-off-by: Hank Donnay <hdonnay@redhat.com>