jar: handle archives with invalid filesystem semantics
Ivan found a jar in the wild that doesn't have filesystem semantics.
Instead of entries `a/` and `a/b`, it contains `a` and `a/b`. This
tricks the zip package into synthesizing a file entry that cannot be
opened.

This change examines the zip file contents directly, instead of using a
filesystem abstraction.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>