dockerfile: handle comments correctly
I consulted the official Dockerfile parser, and this seems more correct.
Initial whitespace is ignored when parsing comments, which is done first
on every line. Dockerfile comments (as opposed to comments in any
embedded language) can only start a line, and are logically deleted from
the token stream. Except they're not because actually ignored because
they can contain parser directives.

The handling of leading spaces continues to be a divergence from the
official parser, because its behavior is bonkers.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>