libindex: fix integration test
AFAICT, this test never worked properly. The updated version of gomock
uses the testing.T's Cleanup method to ensure the controller's Finish
method is called, which was not happening in the test as written. In
addition, because the unused mock structs were being stored in a slice
that was mutated in multiple spots, the "unused variable" compiler and
vet errors didn't trip.

This commit ensures the mock is used, but I uncovered another surprising
fact: cached and live results are different. That's a bigger bug for
another day, though.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>