chore: new exportable indexer and datastore packages
In order to leverage claircore as a library dependency some
packages (and associated interfaces) need to be publicly consumable.
This change also unifies the postgres implementations.

In order to implement a new backend or a noop backend it'd
be good to get all the postgres specific implementation
details in one place.

Signed-off-by: crozzy <joseph.crosland@gmail.com>