rpm: remote rpm packageScanner from rpm ecosystem
Currently the rpm packageScanner is a part of the rpm Ecosystem
and and the rhel Ecosystem, this cause a multiple environments to
be returned for the same package_db. This change could be strange
with regards to defaults as a user would need the rhel Ecosystem to
be required in order to find packages in a suse distro.

Signed-off-by: crozzy <joseph.crosland@gmail.com>