rhel: Check that after casting to mappingFile we have a usable mapper
Currently it is possible that if the repo2cpe_mapping_url or the
repo2cpe_mapping_file (or indeed if the endpoint is down) that we will
panic as the mappingFile will cast to a nil. This will check for a nil
mapper before it gets accessed and error out. This is also an issue
for name2repos_mapping_url and name2repos_mapping_file used by RHCC
scanner.

Signed-off-by: crozzy <joseph.crosland@gmail.com>