rpm: use new bdb package
This removes the extract+exec flow for bdb-based rpm databases.
There are some changes in the package extraction to account for how the
`rpm` command presented output that may cause churn in how the `ndb` and
`sqlite` formats reported. The indexer version is incremented just to be
on the safe side.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>