tests: Rename testdata to .testdata
We recently had an issue where test data file names weren't
compatible with the go tools. That led to the decision to stop any
testdata from being pulled when using claircore as a dep. This should
cut-down on final image size too. NOTE-TO-DELETE: just renamed the dirs
in this commit.

Signed-off-by: crozzy <joseph.crosland@gmail.com>