whiteout: Add the ability to reason about whiteout files in layers
As Clair currently analyses layer files individually there is no
mechanism for reading whiteout files and deciding if their existance
removes a package from the final image. This change adds the concept
of a Resolver that runs after the ecosystem's coalescer's, it gives
the Resolver the full context of what is in each layer. We also
introduce a Whiteout ecosystem (FileScanner and Coalescer)to persist
whiteout files as part of the index report and adds a File field to the
package definition to tell the whiteout Resolver where a package
came from and decern if it has been removed.

Signed-off-by: crozzy <joseph.crosland@gmail.com>