whiteout: Account for null filepath column
For all existing entries in the package_scanartifact table, filepath
will be null. This is not an issue until one of the scanner versions
is updated and the indexing process will attempt to reindex the layers
with the new scanner and retrieve the existing packages from the
package_scanartifact (ie. reconstruct the index_report). This change
ensures null is handled.

Signed-off-by: crozzy <joseph.crosland@gmail.com>