chore: Only try building and pushing golang image when in quay
This change will only attempt to build and push the golang image when
it's being run from quay/claircore, no other fork should attempt this
as it's pushed to particular container repo requiring secrets.
Something to keep in mind is this will silently not trigger if we
migrate to a new namespace.

Signed-off-by: crozzy <joseph.crosland@gmail.com>