indexer: Scope versioned scanners to controller
The indexer.Options are mostly shared objects that are instanciated
during startup (i.e. Layerscanner, FetchArena etc), however
the Vscnrs can be modified during indexing (when re-evaluating which
scanners need to used for a manifest that is partially up to date).
This situation is exacerbated when the same manifest is submitted at
the same time.

Signed-off-by: crozzy <joseph.crosland@gmail.com>