aws: Fix regex for Amazon Linux 1 versions
While the last release of AL1 (formerly Amazon Linux AMI) was
the version "2018.03", there were a number of them before that.

On Dockerhub, it goes all the way back to 2016.09, so we should
be matching the os-release correctly for all of those versions.

Add examples extracted from the container images, and use the
full CPE string to match them.

Signed-off-by: Stewart Smith <trawets@amazon.com>