aws: account for AL2023 mirror.list not having newline
The previous mirror.list endpoints for AL1 and AL2 had newlines at the
end which the code accounted for (by splitting on newlines and
discounting the last entry). AL2023 doesn't have a new line so we were
consuming the one url and then discounting it. This change trims the
final newline if it exists and then splits on newlines.

Signed-off-by: crozzy <joseph.crosland@gmail.com>