rhcc: add a resolver to discount non-rhcc packages
This change adds an rhcc resolver that will check which packages
came from the same layer as a Red Hat container layer and remove
them from the final index report. This is to address the false
positives that can be created when RH patches packages but they
still report their vulnerable versions.

Signed-off-by: crozzy <joseph.crosland@gmail.com>