tarfs: rewrite to support random access archives
This change adds support for some schemes of encoding a table of
contents and allowing individual files to be accessed independently.
It breaks the API by adding a "size" parameter (like the archive/zip
package) but attempts to autodetect when given a nonsense value.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>
Change-Id: Ic0e6758129c3e5632b91cb933d845e7d6a6a6964
JJ: See-Also: CLAIRDEV-NNNN
JJ: Closes: #NNNN