indexer: Reprocess unfinished index reports
If a manifest has scanned with all it's scanners it _should_ end up with a completed index report in the IndexFinished state.  If it does not then it will not be possible to query this manifest.  To fix an incomplete index report it will need rescanning so trigger a rescan.

Signed-off-by: Iain Duncan <iain.duncan@uk.ibm.com>