integration: make `PGVERSION` a pattern
This change makes it so that specifying a pattern like `latest` or `15`
looks at the bill of materials for the embedded PostgreSQL binaries and
fetches a matching version. Previously, this environment variable needed
to be a full `<major>.<minor>.<patch>` version. That's still allowed if
a specific version is needed.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>