zreader: remove zlib support
Noticed on #1213 that the zreader test was failing. I was able to
reproduce reliably by just upping the number of test runs.

Ultimately, the zlib header is just 2 bytes and not sufficiently unique
enough to get the autodetection to work. Currently we don't consume
anything that's "bare" zlib (outside of HTTP requests, which are handled
by `net/http` transparently), so I think it's fine to just drop it.

We never hit a problem in the fetcher because gzip and uncompressed tars
have a sufficiently different header from the zlib header that it
doesn't trip.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>