rhel: update RHEL matcher to account for CPE subset matching
Start matching repository CPEs based on the CPE subset relation.
This change interprets VEX CPEs identifying Red Hat repositories as CPE
matching expressions and looks for a subset relation with the record's
repositoty CPE. This change also introduces a fallback to deal with CPEs
in the VEX data that are expected to describe a subset relationship but
don't use the correct matching syntax, in these cases matching is done
with a crude string prefix match.

Signed-off-by: crozzy <joseph.crosland@gmail.com>