matcher: cache query results for the same query
There is a possiblity that if the index record is referring to the same
package but different repositories and the Query() function isn't adding
a matcher constraint for the repository, queries to the DB can end up
being identical. This patch makes sure that the same query is only
executed once by caching the results and using them for subsequent
calls.

Signed-off-by: crozzy <joseph.crosland@gmail.com>