vex: add non-vuln advisories to the deleted slice
There is a corner-case where an advisory can initially show products as
known_affected then as known_not_affected. Because this updated advisory
doesn't result in vulnerabilities the previous vulnerabilities
associated with this advisory are carried forward. This change adds any
advisories that don't lead to created vulnerabilities to the deleted
slice in-order to ensure no existing vulnerabilities that could be
related to this advisory are carried forward. In essence, it is as-if
the advisory has been parsed from the deletions.csv file.

Signed-off-by: crozzy <joseph.crosland@gmail.com>