migrations: add missing index creating migration
While the sql command was added to claircore (and the admin command was
added to clairctl) to create an index it was never actually reference in
the migrations. This should be a no-op for all instances that ran the
associated clairctl admin command. Also, added a test to try and avoid
this in the future.

Signed-off-by: crozzy <joseph.crosland@gmail.com>