gobin: always report the main version in Package.Version
Go 1.24 changes the way in which the main module's version is reported
when the module is build via `go build`. What used to be just reported
as `(devel)` will now be reported as the version + revision info
(+dirty if status is uncommitted) unless the -buildvcs=false is set. All
this means that the unit tests failed as the expected `(devel)` version
wasn't found and instead "" was reported because the gobin detector
doesn't set any version if the Version was parseable (this isn't a big
deal because the matching always works of NormalizedVersion which is
always set). This change always adds the Version and the test checks the
Version with the same regex `ParseVersion` uses.

Signed-off-by: crozzy <joseph.crosland@gmail.com>