rpm: optimize file checking operation
While load testing the latest version it was found that the indexing
latency was drastically increase (7x was observed). After some
profile/blocking analysis it was found that the volume of calls to the
FileInstalledByRPM() function were causing lock contention due to sheer
number of paths the language detectors were checking. This change
addresses that issue by creating FileChecker object before path
iteration and then using it for the subsequent paths. After this change,
no perceivable latency was added as a result of the RPM file checking.

Signed-off-by: crozzy <joseph.crosland@gmail.com>