rhcc: consider product CPE when creating VEX record
Previously all OCI advisories in VEX were given the same
rhcc.GoldRepo and the package version information was used to infer
product information. With the addition of a labels.json file and a
timestamp version for containers we need to take into account the
product CPE that the VEX surfaces during the matching process.

Signed-off-by: crozzy <joseph.crosland@gmail.com>