rhcc: modify matching to account for CPE matching
With the introduction on the labels.json file to identify RH produced
container images Clair needs to start accounting for the CPE surfaced in
the labels.json file and the VEX data. Care is taken here to ensure the
original Dockerfile parsing process still works as expected (although,
dockerfile images will require a re-index to include needed metadata).

Signed-off-by: crozzy <joseph.crosland@gmail.com>