db: remove redundant updater clause
When updating the enrichment entries there is an association
query that contains a sub-query to look up an enrichment's ID based
on the hash_kind, hash and updater. This query was causing the planner
to prefer the updater index (which is pretty much all the records) and
not the hash_kind, hash index. Because there is already a unique
contraint on hash_kind, hash the updater condition was redundant and
could be removed.

Signed-off-by: crozzy <joseph.crosland@gmail.com>