postgres: add ANALYZE statements to update process
It's possible for the bulk updating to leave the table stats in an
unrepresentative way. This means that subsequent queries (be they
updates, gets or deletes) run very slow and hold table locks for an
unreasonable period of time. This change manually performs ANALYSIS
to keep the table stats up to date.

Signed-off-by: crozzy <joseph.crosland@gmail.com>