rhel: don't add repo id to environment if repo isn't valid
There is a case where the repository information in the layer isn't
valid and the repoid in the DNF DB doesn't refer to a valid Red Hat
repo. In this case we don't associate the package to any repository.
Previously this case would have meant the package would be associated to
all repositories.

Signed-off-by: crozzy <joseph.crosland@gmail.com>