rhel: update repository detector to utilize URI field
Given that there is a unique index on the repository table (name, key, URI) and
considering the fact that the name was being used to hold the repoid (in the
case of DNF) there was never a chance to save more than one repo for each repoid
(which can be possible when considering the repository_to_cpe mapping file).
This patch utilized the URI field to carry the repoid metadata information and
makes the repo.Name consistent (an unbound CPE).

Signed-off-by: crozzy <joseph.crosland@gmail.com>