internal/cache: improve test
This loops calling both `runtime.GC` and `runtime.Gosched` to trigger
garbage collection and the running of the cleanup functions. Instead of
running a set number of times, it runs for a bounded amount of time.

On my machine, this passes with `-count=1000` which is about as good as
I think it can get. The previous version would reliably fail with counts
as low as 10.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>