rpm: close sqlite resources on the error path
It was observed in production that there was a goroutine leak, this was
traced back to a database/sql.(*DB).connectionOpener+0x86 call and the
rpm sqlite support. There is a case where the DB is open and the Ping()
fails. This means there is no Close() issued, no spool file clean up and
no finalizer supported added, so it's not caught in a panic.

Signed-off-by: crozzy <joseph.crosland@gmail.com>